+++ src/main/java/net/sf/jabref/gui/BasePanel.java
@@ -98,6 +98,7 @@ import net.sf.jabref.logic.autocompleter.AutoCompleter;
 import net.sf.jabref.logic.autocompleter.AutoCompleterFactory;
 import net.sf.jabref.logic.autocompleter.ContentAutoCompleters;
 import net.sf.jabref.logic.bibtexkeypattern.BibtexKeyPatternUtil;
+import net.sf.jabref.logic.citationstyle.CitationStyleCache;
 import net.sf.jabref.logic.exporter.BibtexDatabaseWriter;
 import net.sf.jabref.logic.exporter.FileSaveSession;
 import net.sf.jabref.logic.exporter.SaveException;
@@ -124,6 +125,7 @@ import net.sf.jabref.model.event.EntryAddedEvent;
 import net.sf.jabref.model.event.EntryChangedEvent;
 import net.sf.jabref.preferences.HighlightMatchingGroupPreferences;
 import net.sf.jabref.preferences.JabRefPreferences;
+import net.sf.jabref.preferences.PreviewPreferences;
 import net.sf.jabref.shared.DBMSSynchronizer;
 import net.sf.jabref.specialfields.Printed;
 import net.sf.jabref.specialfields.Priority;
@@ -150,6 +152,8 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
 
     private final BibDatabaseContext bibDatabaseContext;
     private final MainTableDataModel tableModel;
+
+    private final CitationStyleCache citationStyleCache;
 
     // To contain instantiated entry editors. This is to save time
     // As most enums, this must not be null
@@ -218,6 +222,8 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
         this.sidePaneManager = frame.getSidePaneManager();
         this.frame = frame;
         this.tableModel = new MainTableDataModel(getBibDatabaseContext());
+
+        citationStyleCache = new CitationStyleCache(bibDatabaseContext);
 
         searchBar = new SearchBar(this);
 
@@ -691,8 +697,9 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
         }
 
         actions.put(Actions.TOGGLE_PREVIEW, (BaseAction) () -> {
-            boolean enabled = !Globals.prefs.getBoolean(JabRefPreferences.PREVIEW_ENABLED);
-            Globals.prefs.putBoolean(JabRefPreferences.PREVIEW_ENABLED, enabled);
+            PreviewPreferences previewPreferences = new PreviewPreferences(Globals.prefs);
+            boolean enabled = !previewPreferences.isPreviewEnabled();
+            previewPreferences.setPreviewEnabled(enabled);
             setPreviewActiveBasePanels(enabled);
             frame.setPreviewToggle(enabled);
         });
@@ -715,7 +722,8 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
             groupsHighlightListener.listChanged(null);
         });
 
-        actions.put(Actions.SWITCH_PREVIEW, (BaseAction) selectionListener::switchPreview);
+        actions.put(Actions.NEXT_PREVIEW_STYLE, (BaseAction) selectionListener::nextPreviewStyle);
+        actions.put(Actions.PREVIOUS_PREVIEW_STYLE, (BaseAction) selectionListener::previousPreviewStyle);
 
         actions.put(Actions.MANAGE_SELECTORS, (BaseAction) () -> {
             ContentSelectorDialog2 csd = new ContentSelectorDialog2(frame, frame, BasePanel.this, false, null);
@@ -1545,7 +1553,7 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
     public void adjustSplitter() {
         if (mode == BasePanelMode.SHOWING_PREVIEW) {
             splitPane.setDividerLocation(
-                    splitPane.getHeight() - Globals.prefs.getInt(JabRefPreferences.PREVIEW_PANEL_HEIGHT));
+                    splitPane.getHeight() - new PreviewPreferences(Globals.prefs).getPreviewPanelHeight());
         } else {
             splitPane.setDividerLocation(
                     splitPane.getHeight() - Globals.prefs.getInt(JabRefPreferences.ENTRY_EDITOR_HEIGHT));
@@ -1667,8 +1675,8 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
             Globals.prefs.putInt(JabRefPreferences.ENTRY_EDITOR_HEIGHT,
                     splitPane.getHeight() - splitPane.getDividerLocation());
         } else if (mode == BasePanelMode.SHOWING_PREVIEW) {
-            Globals.prefs.putInt(JabRefPreferences.PREVIEW_PANEL_HEIGHT,
-                    splitPane.getHeight() - splitPane.getDividerLocation());
+            int previewPanelHeight = splitPane.getHeight() - splitPane.getDividerLocation();
+            new PreviewPreferences(Globals.prefs).setPreviewPanelHeight(previewPanelHeight);
         }
         mode = BasePanelMode.SHOWING_EDITOR;
         currentEditor = editor;
@@ -1953,8 +1961,8 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
      */
     public void saveDividerLocation() {
         if (mode == BasePanelMode.SHOWING_PREVIEW) {
-            Globals.prefs.putInt(JabRefPreferences.PREVIEW_PANEL_HEIGHT,
-                    splitPane.getHeight() - splitPane.getDividerLocation());
+            int previewPanelHeight = splitPane.getHeight() - splitPane.getDividerLocation();
+            new PreviewPreferences(Globals.prefs).setPreviewPanelHeight(previewPanelHeight);
         } else if (mode == BasePanelMode.SHOWING_EDITOR) {
             Globals.prefs.putInt(JabRefPreferences.ENTRY_EDITOR_HEIGHT,
                     splitPane.getHeight() - splitPane.getDividerLocation());
@@ -2414,4 +2422,13 @@ public class BasePanel extends JPanel implements ClipboardOwner, FileUpdateListe
     public BibDatabaseContext getDatabaseContext() {
         return bibDatabaseContext;
     }
+
+    public CitationStyleCache getCitationStyleCache() {
+        return citationStyleCache;
+    }
+
+    public PreviewPanel getPreviewPanel() {
+        return currentPreview;
+    }
+
 }