+++ src/main/java/net/sf/jabref/gui/JabRefFrame.java
@@ -125,6 +125,7 @@ import net.sf.jabref.model.entry.EntryType;
 import net.sf.jabref.preferences.HighlightMatchingGroupPreferences;
 import net.sf.jabref.preferences.JabRefPreferences;
 import net.sf.jabref.preferences.LastFocusedTabPreferences;
+import net.sf.jabref.preferences.PreviewPreferences;
 import net.sf.jabref.specialfields.Printed;
 import net.sf.jabref.specialfields.Priority;
 import net.sf.jabref.specialfields.Quality;
@@ -369,9 +370,12 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
             Localization.lang("Disable highlight groups matching entries")));
 
 
-    private final AbstractAction switchPreview = new GeneralAction(Actions.SWITCH_PREVIEW,
-            Localization.menuTitle("Switch preview layout"),
-            Globals.getKeyPrefs().getKey(KeyBinding.SWITCH_PREVIEW_LAYOUT));
+    private final AbstractAction nextPreviewStyle = new GeneralAction(Actions.NEXT_PREVIEW_STYLE,
+            Localization.menuTitle("Next preview layout"),
+            Globals.getKeyPrefs().getKey(KeyBinding.NEXT_PREVIEW_LAYOUT));
+    private final AbstractAction previousPreviewStyle = new GeneralAction(Actions.PREVIOUS_PREVIEW_STYLE,
+            Localization.menuTitle("Previous preview layout"),
+            Globals.getKeyPrefs().getKey(KeyBinding.PREVIOUS_PREVIEW_LAYOUT));
     private final AbstractAction makeKeyAction = new GeneralAction(Actions.MAKE_KEY,
             Localization.menuTitle("Autogenerate BibTeX keys"),
             Localization.lang("Autogenerate BibTeX keys"),
@@ -638,8 +642,12 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
                 return;
             }
 
+            if (bp.getPreviewPanel() != null) {
+                bp.getPreviewPanel().updateLayout();
+            }
+
             groupToggle.setSelected(sidePaneManager.isComponentVisible("groups"));
-            previewToggle.setSelected(Globals.prefs.getBoolean(JabRefPreferences.PREVIEW_ENABLED));
+            previewToggle.setSelected(new PreviewPreferences(Globals.prefs).isPreviewEnabled());
             fetcherToggle.setSelected(sidePaneManager.isComponentVisible(generalFetcher.getTitle()));
             Globals.getFocusListener().setFocused(bp.getMainTable());
             setWindowTitle();
@@ -1278,7 +1286,8 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
         view.add(new JCheckBoxMenuItem(enableToggle(generalFetcher.getAction())));
         view.add(new JCheckBoxMenuItem(toggleGroups));
         view.add(new JCheckBoxMenuItem(togglePreview));
-        view.add(getSwitchPreviewAction());
+        view.add(getNextPreviewStyleAction());
+        view.add(getPreviousPreviewStyleAction());
 
         mb.add(view);
 
@@ -1516,7 +1525,7 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
                 sendAsEmail, downloadFullText, writeXmpAction, optMenuItem, findUnlinkedFiles, addToGroup, removeFromGroup,
                 moveToGroup, autoLinkFile, resolveDuplicateKeys, openUrl, openFolder, openFile, togglePreview,
                 dupliCheck, autoSetFile, newEntryAction, newSpec, customizeAction, plainTextImport, getMassSetField(), getManageKeywords(),
-                pushExternalButton.getMenuAction(), closeDatabaseAction, getSwitchPreviewAction(), checkIntegrity,
+                pushExternalButton.getMenuAction(), closeDatabaseAction, getNextPreviewStyleAction(), getPreviousPreviewStyleAction(), checkIntegrity,
                 toggleHighlightAny, toggleHighlightAll, toggleHighlightDisable, databaseProperties, abbreviateIso, abbreviateMedline,
                 unabbreviate, exportAll, exportSelected, importCurrent, saveAll, focusTable, increaseFontSize, decreseFontSize,
                 toggleRelevance, toggleQualityAssured, togglePrinted, pushExternalButton.getComponent()));
@@ -2306,8 +2315,12 @@ public class JabRefFrame extends JFrame implements OutputPrinter {
         return back;
     }
 
-    public AbstractAction getSwitchPreviewAction() {
-        return switchPreview;
+    public AbstractAction getNextPreviewStyleAction() {
+        return nextPreviewStyle;
+    }
+
+    public AbstractAction getPreviousPreviewStyleAction() {
+        return previousPreviewStyle;
     }
 
     public JSplitPane getSplitPane() {