+++ src/main/java/net/sf/jabref/gui/PreviewPanel.java
@@ -8,7 +8,6 @@ import java.awt.print.PrinterException;
 import java.io.IOException;
 import java.io.StringReader;
 import java.lang.reflect.InvocationTargetException;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.regex.Pattern;
 
@@ -35,6 +34,8 @@ import net.sf.jabref.JabRefExecutorService;
 import net.sf.jabref.gui.desktop.JabRefDesktop;
 import net.sf.jabref.gui.fieldeditors.PreviewPanelTransferHandler;
 import net.sf.jabref.gui.keyboard.KeyBinding;
+import net.sf.jabref.gui.worker.CitationStyleWorker;
+import net.sf.jabref.logic.citationstyle.CitationStyle;
 import net.sf.jabref.logic.exporter.ExportFormats;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.layout.Layout;
@@ -43,6 +44,7 @@ import net.sf.jabref.logic.search.SearchQueryHighlightListener;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.event.FieldChangedEvent;
+import net.sf.jabref.preferences.PreviewPreferences;
 
 import com.google.common.eventbus.Subscribe;
 import org.apache.commons.logging.Log;
@@ -51,8 +53,7 @@ import org.apache.commons.logging.LogFactory;
 /**
  * Displays an BibEntry using the given layout format.
  */
-public class PreviewPanel extends JPanel
-        implements SearchQueryHighlightListener, EntryContainer {
+public class PreviewPanel extends JPanel implements SearchQueryHighlightListener, EntryContainer {
 
     private static final Log LOGGER = LogFactory.getLog(PreviewPanel.class);
 
@@ -67,26 +68,20 @@ public class PreviewPanel extends JPanel
      */
     private Optional<BibDatabaseContext> databaseContext = Optional.empty();
 
+    private Optional<BasePanel> basePanel = Optional.empty();
+
+    private boolean fixedLayout;
     private Optional<Layout> layout = Optional.empty();
-
-    /**
-     * must not be null, must always be set during constructor, but can change over time
-     */
-    private String layoutFile;
-
-    private final Optional<BasePanel> basePanel;
-
     private JEditorPane previewPane;
 
     private final JScrollPane scrollPane;
 
-    private final PrintAction printAction;
-
-    private final CloseAction closeAction;
-
-    private final CopyPreviewAction copyPreviewAction;
+    private final PrintAction printAction = new PrintAction();
+    private final CloseAction closeAction = new CloseAction();
+    private final CopyPreviewAction copyPreviewAction = new CopyPreviewAction();
 
     private Optional<Pattern> highlightPattern = Optional.empty();
+    private CitationStyleWorker citationStyleWorker;
 
     /**
      * @param databaseContext
@@ -97,12 +92,9 @@ public class PreviewPanel extends JPanel
      * @param panel
      *            (may be null) If not given no toolbar is shown on the right
      *            hand side.
-     * @param layoutFile
-     *            (must be given) Used for layout
      */
-    public PreviewPanel(BibDatabaseContext databaseContext, BibEntry entry,
-                        BasePanel panel, String layoutFile) {
-        this(panel, databaseContext, layoutFile);
+    public PreviewPanel(BibDatabaseContext databaseContext, BibEntry entry, BasePanel panel) {
+        this(panel, databaseContext);
         setEntry(entry);
     }
 
@@ -113,23 +105,15 @@ public class PreviewPanel extends JPanel
      *            hand side.
      * @param databaseContext
      *            (may be null) Used for resolving pdf directories for links.
-     * @param layoutFile
-     *            (must be given) Used for layout
      */
-    public PreviewPanel(BasePanel panel, BibDatabaseContext databaseContext, String layoutFile) {
+    public PreviewPanel(BasePanel panel, BibDatabaseContext databaseContext) {
         super(new BorderLayout(), true);
 
         this.databaseContext = Optional.ofNullable(databaseContext);
-        this.layoutFile = Objects.requireNonNull(layoutFile);
-        updateLayout();
-
-        this.closeAction = new CloseAction();
-        this.printAction = new PrintAction();
-        this.copyPreviewAction = new CopyPreviewAction();
-
         this.basePanel = Optional.ofNullable(panel);
 
         createPreviewPane();
+        updateLayout();
 
         if (panel != null) {
             // dropped files handler only created for main window
@@ -142,7 +126,6 @@ public class PreviewPanel extends JPanel
                 ScrollPaneConstants.VERTICAL_SCROLLBAR_AS_NEEDED,
                 ScrollPaneConstants.HORIZONTAL_SCROLLBAR_NEVER);
         scrollPane.setBorder(null);
-
 
         add(scrollPane, BorderLayout.CENTER);
 
@@ -166,7 +149,8 @@ public class PreviewPanel extends JPanel
         JPopupMenu menu = new JPopupMenu();
         menu.add(this.printAction);
         menu.add(this.copyPreviewAction);
-        this.basePanel.ifPresent(p -> menu.add(p.frame().getSwitchPreviewAction()));
+        this.basePanel.ifPresent(p -> menu.add(p.frame().getNextPreviewStyleAction()));
+        this.basePanel.ifPresent(p -> menu.add(p.frame().getPreviousPreviewStyleAction()));
         return menu;
     }
 
@@ -203,12 +187,43 @@ public class PreviewPanel extends JPanel
         this.databaseContext = Optional.ofNullable(databaseContext);
     }
 
-    public void updateLayout(String layoutFormat) {
-        layoutFile = layoutFormat;
-        updateLayout();
+    public BasePanel getBasePanel() {
+        return this.basePanel.orElse(null);
     }
 
-    private void updateLayout() {
+    public void setBasePanel(BasePanel basePanel) {
+        this.basePanel = Optional.ofNullable(basePanel);
+    }
+
+    public void updateLayout() {
+        if (fixedLayout) {
+            LOGGER.debug("cannot change the layout because the layout is fixed");
+            return;
+        }
+
+        PreviewPreferences previewPreferences = new PreviewPreferences(Globals.prefs);
+        String style = previewPreferences.getPreviewCycle().get(previewPreferences.getCyclePreviewPosition());
+
+        if (CitationStyle.isCitationStyleFile(style)) {
+            if (getBasePanel() != null) {
+                layout = Optional.empty();
+                CitationStyle citationStyle = CitationStyle.createCitationStyleFromFile(style);
+                if (citationStyle != null) {
+                    getBasePanel().getCitationStyleCache().setCitationStyle(citationStyle);
+                    getBasePanel().output(Localization.lang("Preview style changed to: %0", citationStyle.getTitle()));
+                }
+            }
+        } else {
+            updatePreviewLayout(previewPreferences.getPreviewStyle());
+            if (getBasePanel() != null) {
+                getBasePanel().output(Localization.lang("Preview style changed to: %0", Localization.lang("Preview")));
+            }
+        }
+
+        update();
+    }
+
+    private void updatePreviewLayout(String layoutFile){
         StringReader sr = new StringReader(layoutFile.replace("__NEWLINE__", "\n"));
         try {
             layout = Optional.of(
@@ -230,7 +245,6 @@ public class PreviewPanel extends JPanel
         bibEntry = Optional.ofNullable(newEntry);
         bibEntry.ifPresent(e -> e.registerListener(this));
 
-        updateLayout();
         update();
     }
 
@@ -250,30 +264,40 @@ public class PreviewPanel extends JPanel
     }
 
     public void update() {
-        StringBuilder sb = new StringBuilder();
         ExportFormats.entryNumber = 1; // Set entry number in case that is included in the preview layout.
-        bibEntry.ifPresent(entry ->
-                layout.ifPresent(acutalLayout -> sb.append(acutalLayout
-                        .doLayout(entry, databaseContext.map(BibDatabaseContext::getDatabase).orElse(null),
-                                highlightPattern)))
-        );
-        String newValue = sb.toString();
 
+        if (citationStyleWorker != null){
+            citationStyleWorker.cancel(true);
+            citationStyleWorker = null;
+        }
+
+        if (layout.isPresent()){
+            StringBuilder sb = new StringBuilder();
+            bibEntry.ifPresent(entry -> sb.append(layout.get()
+                    .doLayout(entry, databaseContext.map(BibDatabaseContext::getDatabase).orElse(null), highlightPattern)));
+            setPreviewLabel(sb.toString());
+        }
+        else if (basePanel.isPresent()){
+            citationStyleWorker = new CitationStyleWorker(this, previewPane);
+            citationStyleWorker.execute();
+        }
+    }
+
+    public void setPreviewLabel(String text) {
         if (SwingUtilities.isEventDispatchThread()) {
-            previewPane.setText(newValue);
+            previewPane.setText(text);
             previewPane.revalidate();
         } else {
             try {
                 SwingUtilities.invokeAndWait(() -> {
-                    previewPane.setText(newValue);
+                    previewPane.setText(text);
                     previewPane.revalidate();
                 });
             } catch (InvocationTargetException | InterruptedException e) {
                 LOGGER.info("Problem setting preview text", e);
             }
         }
-        // Scroll to top:
-        scrollToTop();
+        this.scrollToTop();
     }
 
     private void scrollToTop() {
@@ -284,6 +308,29 @@ public class PreviewPanel extends JPanel
     public void highlightPattern(Optional<Pattern> newPattern) {
         this.highlightPattern = newPattern;
         update();
+    }
+
+    public Optional<Pattern> getHighlightPattern() {
+        return highlightPattern;
+    }
+
+    /**
+     * this fixes the Layout, the user cannot change it anymore. Useful for testing the styles in the settings
+     * @param parameter should be either a {@link String} (for the old PreviewStyle) or a {@link CitationStyle}.
+     */
+    public PreviewPanel setFixedLayout(Object parameter) {
+        this.fixedLayout = true;
+
+        if (parameter instanceof String) {
+            updatePreviewLayout((String) parameter);
+        } else if (parameter instanceof CitationStyle) {
+            layout = Optional.empty();
+            getBasePanel().getCitationStyleCache().setCitationStyle((CitationStyle) parameter);
+        } else {
+            LOGGER.error("unknown style type");
+        }
+        update();
+        return this;
     }
 
     class PrintAction extends AbstractAction {