+++ src/main/java/net/sf/jabref/gui/maintable/MainTableSelectionListener.java
@@ -37,7 +37,7 @@ import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.util.OS;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
-import net.sf.jabref.preferences.JabRefPreferences;
+import net.sf.jabref.preferences.PreviewPreferences;
 import net.sf.jabref.specialfields.SpecialFieldValue;
 import net.sf.jabref.specialfields.SpecialFieldsUtils;
 
@@ -54,14 +54,12 @@ import org.apache.commons.logging.LogFactory;
 public class MainTableSelectionListener implements ListEventListener<BibEntry>, MouseListener,
         KeyListener, FocusListener {
 
-    private final PreviewPanel[] previewPanel;
     private final MainTable table;
     private final BasePanel panel;
     private final EventList<BibEntry> tableRows;
 
-    private int activePreview = Globals.prefs.getInt(JabRefPreferences.ACTIVE_PREVIEW);
     private PreviewPanel preview;
-    private boolean previewActive = Globals.prefs.getBoolean(JabRefPreferences.PREVIEW_ENABLED);
+    private boolean previewActive = new PreviewPreferences(Globals.prefs).isPreviewEnabled();
     private boolean workingOnPreview;
 
     private boolean enabled = true;
@@ -79,25 +77,12 @@ public class MainTableSelectionListener implements ListEventListener<BibEntry>, 
         this.table = table;
         this.panel = panel;
         this.tableRows = table.getTableModel().getTableRows();
-        previewPanel = new PreviewPanel[] {
-                new PreviewPanel(panel.getBibDatabaseContext(), null, panel,
-                        Globals.prefs.get(JabRefPreferences.PREVIEW_0)),
-                new PreviewPanel(panel.getBibDatabaseContext(), null, panel,
-                        Globals.prefs.get(JabRefPreferences.PREVIEW_1))};
-
-        panel.getSearchBar().getSearchQueryHighlightObservable().addSearchListener(previewPanel[0]);
-        panel.getSearchBar().getSearchQueryHighlightObservable().addSearchListener(previewPanel[1]);
-
-        this.preview = previewPanel[activePreview];
+        preview = new PreviewPanel(panel.getBibDatabaseContext(), null, panel);
+        panel.getSearchBar().getSearchQueryHighlightObservable().addSearchListener(preview);
     }
 
     public void setEnabled(boolean enabled) {
         this.enabled = enabled;
-    }
-
-    public void updatePreviews() {
-        previewPanel[0].updateLayout(Globals.prefs.get(JabRefPreferences.PREVIEW_0));
-        previewPanel[1].updateLayout(Globals.prefs.get(JabRefPreferences.PREVIEW_1));
     }
 
     @Override
@@ -458,19 +443,25 @@ public class MainTableSelectionListener implements ListEventListener<BibEntry>, 
         }
     }
 
-    public void switchPreview() {
-        if (activePreview < (previewPanel.length - 1)) {
-            activePreview++;
-        } else {
-            activePreview = 0;
-        }
-        Globals.prefs.putInt(JabRefPreferences.ACTIVE_PREVIEW, activePreview);
-        if (previewActive) {
-            this.preview = previewPanel[activePreview];
+    public void nextPreviewStyle(){
+        cyclePreview(new PreviewPreferences(Globals.prefs).getCyclePreviewPosition() + 1);
+    }
 
-            if (!table.getSelected().isEmpty()) {
-                updatePreview(table.getSelected().get(0), true);
-            }
+    public void previousPreviewStyle(){
+        cyclePreview(new PreviewPreferences(Globals.prefs).getCyclePreviewPosition() - 1);
+    }
+
+    private void cyclePreview(int newPosition) {
+        if (!previewActive){
+            return;
+        }
+
+        new PreviewPreferences(Globals.prefs).setCyclePreviewPosition(newPosition);
+
+        preview.updateLayout();
+        preview.update();
+        if (!table.getSelected().isEmpty()) {
+            updatePreview(table.getSelected().get(0), true);
         }
     }