+++ src/main/java/net/sf/jabref/gui/openoffice/StyleSelectDialog.java
@@ -134,7 +134,7 @@ class StyleSelectDialog {
 
         // Create a preview panel for previewing styles
         // Must be done before creating the table to avoid NPEs
-        preview = new PreviewPanel(null, null, "");
+        preview = new PreviewPanel(null, null);
         // Use the test entry from the Preview settings tab in Preferences:
         preview.setEntry(prevEntry);